requirements.txt
# ============================================================================
# MAX Dependency Counseling Bot - Requirements
# ============================================================================
# Проект: Бот психологической поддержки для людей с зависимостями
# Платформа: MAX Messenger
# Python версия: 3.9+
# ============================================================================

# Асинхронные HTTP запросы для работы с MAX API
aiohttp==3.13.2
    # Асинхронный HTTP клиент/сервер
    # Используется для взаимодействия с MAX Messenger API
    # Обновлено с 3.9.1 до 3.13.2 для стабильности и безопасности

# Управление переменными окружения
python-dotenv==1.0.0
    # Загрузка переменных окружения из .env файла
    # Используется для хранения токенов и конфигурации

# ============================================================================
# Описание архитектуры проекта
# ============================================================================
# 
# СТРУКТУРА ПРОЕКТА:
# ├── main_max.py              - Точка входа для MAX мессенджера
# ├── config.py                - Конфигурация (токены, URL API)
# ├── bot/
# │   ├── max_adapter.py       - Адаптер MAX API (MaxBot, MaxUpdate, MaxMessageProxy)
# │   ├── handlers.py          - Обработчики всех состояний диалога
# │   ├── conversation_flow.py - Генераторы клавиатур и логика навигации
# │   ├── dependency_links.py  - База ссылок на группы поддержки (25 городов)
# │   ├── states.py            - Enum состояний бота
# │   └── utils.py             - Вспомогательные функции (логирование)
#
# ОСНОВНЫЕ КОМПОНЕНТЫ:
#
# 1. MAX API Adapter (bot/max_adapter.py):
#    - MaxBot: HTTP клиент для MAX API
#    - MaxUpdate: Обработка обновлений (message_created, message_callback, bot_started)
#    - MaxMessageProxy: Прокси для отправки сообщений
#    - MaxCallbackQueryProxy: Обработка callback кнопок
#    - Фильтрация дубликатов по timestamp
#
# 2. Conversation Flow (bot/handlers.py):
#    - Приветствие с информацией о зависимостях
#    - Выбор типа зависимости (11 типов)
#    - Выбор часового пояса (11 поясов)
#    - Выбор города (25 городов России)
#    - Выбор типа помощи (группы, специалист, литература)
#    - Подбор специалиста (пол, возраст)
#    - Информационные материалы
#    - Обработка анонимных вопросов
#    - FAQ и вебинары
#
# 3. Dependency Links (bot/dependency_links.py):
#    - Полная база ссылок на группы поддержки
#    - 25 городов: Москва, СПб, Воронеж, Краснодар, Казань и др.
#    - 11 типов зависимостей: алкоголь, наркотики, игры, еда, интернет и др.
#    - Реальные URL ресурсов помощи
#
# 4. Navigation System:
#    - Глобальные обработчики (работают из любого состояния)
#    - Кнопки "Назад" на всех уровнях
#    - Restart conversation
#    - FAQ и вебинары в финальном меню
#
# ОСОБЕННОСТИ РЕАЛИЗАЦИИ:
#
# 1. Long Polling:
#    - Marker-based пагинация для получения обновлений
#    - Timestamp filtering для предотвращения дубликатов
#    - Обработка событий: message_created, message_callback, bot_started
#
# 2. State Management:
#    - Хранение состояния пользователя в памяти (user_states)
#    - Контекст пользователя (user_data) для сохранения выборов
#    - Поддержка множества одновременных пользователей
#
# 3. Автоматическое приветствие:
#    - Обработка события bot_started
#    - Автоматический старт для новых пользователей
#    - Виртуальное сообщение для reply_text
#
# 4. Обработка ошибок:
#    - Try-catch обертки для всех обработчиков
#    - Детальное логирование всех операций
#    - Graceful degradation при ошибках API
#
# КОНФИГУРАЦИЯ:
#
# Переменные окружения (.env):
# - BOT_TOKEN: Токен авторизации MAX бота
# - MAX_API_BASE_URL: https://platform-api.max.ru
#
# MAX API Authentication:
# - Authorization header с токеном
# - User ID: 100371934 (Роман Сурмач)
# - Bot ID: 71638156 (@t38_hakaton_bot)
#
# ЗАВИСИМОСТИ ПО ТИПАМ:
#
# 1. Алкогольная зависимость (alcohol)
# 2. Наркотическая зависимость (drugs)
# 3. Игровая зависимость (gaming)
# 4. Пищевая зависимость (food)
# 5. Интернет-зависимость (internet)
# 6. Никотиновая зависимость (nicotine)
# 7. Созависимость (codependency)
# 8. ВСД/ПА зависимость (vad)
# 9. Любовная зависимость (love)
# 10. Трудоголизм (workaholism)
# 11. Виртуальные отношения (vr)
#
# ГОРОДА (25):
# Москва, Санкт-Петербург, Воронеж, Краснодар, Казань, Самара, Ижевск,
# Екатеринбург, Челябинск, Омск, Барнаул, Новосибирск, Красноярск, Иркутск,
# Улан-Удэ, Якутск, Благовещенск, Владивосток, Хабаровск, Магадан,
# Южно-Сахалинск, Петропавловск-Камчатский, Анадырь, Калининград
#
# ============================================================================
# Установка и запуск
# ============================================================================
#
# 1. Установка зависимостей:
#    pip install -r requirements.txt
#
# 2. Настройка переменных окружения:
#    Создать файл .env с:
#    BOT_TOKEN=<ваш_токен_MAX_бота>
#    MAX_API_BASE_URL=https://platform-api.max.ru
#
# 3. Запуск бота:
#    python main_max.py
#
# 4. Альтернативный запуск (Windows):
#    run_max_bot.bat
#
# ============================================================================
# Версионирование
# ============================================================================
#   - MAX Messenger Adapter (текущая версия)
#   - Полная адаптация под MAX API
#   - Long polling с фильтрацией дубликатов
#   - Обработка bot_started событий
#   - 25 городов × 11 зависимостей
#   - Автоматическое приветствие
#   - Навигация "Назад" на всех уровнях
#   - FAQ и вебинары
# ============================================================================